Конечная, сколько всего было переделано, доделано, обновлено, убрано

+++ HealthPatient/Models/Review.cs
@@ -11,7 +11,7 @@ public partial class Review
 
     public int? DoctorId { get; set; }
 
-    public short? Rating { get; set; }
+    public int? Rating { get; set; }
 
     public string? Comment { get; set; }
